+++ packages/webawesome/src/components/select/select.ts
@@ -83,6 +83,7 @@ import styles from './select.styles.js';
  * @cssproperty [--tag-max-size=10ch] - When using `multiple`, the max size of tags before their content is truncated.
  *
  * @cssstate blank - The select is empty.
+ * @cssstate focused - The select has focus.
  */
 @customElement('wa-select')
 export default class WaSelect extends WebAwesomeFormAssociatedElement {